Run bench/bsbm_ with set -eu to exit right away if something abnormal happens
e.g. if one command fails or undefined variable is used.

In turn it would allow to address such codeshell warnings like

	In bench/bsbm_graphdb.sh line 7:
	cd bsbm-tools
	^-----------^ SC2164 (warning): Use 'cd ... || exit' or 'cd ... || return' in case cd fails.

	Did you mean:
	cd bsbm-tools || exit

but note that there are uses left like

	❯ git grep 'cd bsbm-tools || exit'
	bench/bsbm_blazegraph.sh:cd bsbm-tools || exit
	bench/bsbm_jena.sh:cd bsbm-tools || exit
	bench/bsbm_rdf4j.sh:cd bsbm-tools || exit

but may be `|| exit` should be removed and they should exit with error? or the
same should be added to the other 3?

	❯ git grep 'cd bsbm-tools$'
	bench/bsbm_graphdb.sh:cd bsbm-tools
	bench/bsbm_oxigraph.sh:cd bsbm-tools
	bench/bsbm_virtuoso.sh:cd bsbm-tools